requirements.txt
openai==1.11.1
langchain==0.1.13
langchain-openai==0.0.8
httpx==0.27.0
PyJWT==2.8.0
faiss-cpu
python-dotenv
pydantic==1.10.12
uvicorn
fastapi
PyMuPDF
PyPDF2
pathlib
python-multipart
tiktoken